pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cxagent"
version = "0.1.0"
description = "SAP Commerce Cloud AI Shopping Agent"
readme = "README.md"
license = "LicenseRef-BSL-1.1"
license-files = ["LICENSE"]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]

dependencies = [
    # Core API
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.41.0",
    "python-dotenv>=1.2.0",
    "python-multipart>=0.0.22",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.13.0",
    # LangChain / LangGraph
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langgraph>=1.1.0",
    "langgraph-checkpoint>=4.0.0",
    "langgraph-prebuilt>=1.0.0",
    "langgraph-sdk>=0.3.0",
    # Anthropic SDK
    "anthropic>=0.84.0",
    # MCP
    "mcp>=1.26.0",
    "fastmcp>=3.1.0",
    # HTTP / SAP OCC
    "httpx>=0.28.0",
    "httpx-sse>=0.4.0",
    # Data / Serialisation
    "orjson>=3.11.0",
    "pyyaml>=6.0.0",
    # Observability
    "langsmith>=0.7.0",
    "opentelemetry-api>=1.40.0",
    # Websockets / SSE
    "websockets>=16.0",
    "sse-starlette>=3.3.0",
    # Security
    "cryptography>=46.0.0",
    "PyJWT>=2.12.0",
    "Authlib>=1.6.0",
    # Utilities
    "tenacity>=9.1.0",
    "requests>=2.32.0",
    # Recommendations (vector math)
    "numpy>=1.26.0",
]

[project.optional-dependencies]
redis = ["redis>=7.3.0"]
payments = ["stripe>=11.0.0"]
search = ["qdrant-client>=1.9.0", "fastembed>=0.7.0"]
gemini = ["langchain-google-genai>=2.1.0,<3.0.0"]
ollama = ["langchain-ollama>=1.0.0", "ollama>=0.6.0"]
# Image search: CLIP model for visual product matching
vision = ["transformers>=4.40.0", "torch>=2.2.0", "Pillow>=10.0.0"]
# Audio search: Whisper model for speech-to-text
audio = ["transformers>=4.40.0", "torch>=2.2.0", "librosa>=0.10.0", "soundfile>=0.12.0"]
# All AI features (recommendations + image + audio)
ai-features = ["cxagent[search,vision,audio]", "numpy>=1.26.0"]
all = ["cxagent[redis,payments,search,gemini,ollama,vision,audio]"]
dev = ["pytest>=9.0.0", "pytest-asyncio>=1.3.0"]

[project.scripts]
cxagent = "main:main"

[tool.setuptools]
packages = { find = { include = ["app*", "acp*", "impex*"] } }
py-modules = ["main", "sap_commerce_mcp_server", "token_vault", "load_products_to_qdrant", "reset_and_load_qdrant"]

[tool.setuptools.package-data]
app = ["static/**/*"]
